Move configuration role integration tests into their own test class
Those tests didn't have anything to do with configuration attributes, so they have been moved
into their own class.
