+++ subprojects/dependency-management/src/integTest/groovy/org/gradle/integtests/resolve/ConfigurationAttributesResolveIntegrationTest.groovy
@@ -18,7 +18,6 @@ package org.gradle.integtests.resolve
 import org.gradle.integtests.fixtures.AbstractIntegrationSpec
 import org.gradle.integtests.fixtures.FluidDependenciesResolveRunner
 import org.junit.runner.RunWith
-import spock.lang.Unroll
 
 @RunWith(FluidDependenciesResolveRunner)
 class ConfigurationAttributesResolveIntegrationTest extends AbstractIntegrationSpec {
@@ -1122,144 +1121,4 @@ class ConfigurationAttributesResolveIntegrationTest extends AbstractIntegrationS
         notExecuted ':b:fooJar'
     }
 
-    @Unroll("cannot resolve a configuration with role #role at execution time")
-    def "cannot resolve a configuration which is for publishing only at execution time"() {
-        given:
-        buildFile << """
-
-        configurations {
-            internal {
-                $code
-            }
-        }
-        dependencies {
-            internal files('foo.jar')
-        }
-
-        task checkState {
-            doLast {
-                configurations.internal.resolve()
-            }
-        }
-
-        """
-
-        when:
-        fails 'checkState'
-
-        then:
-        failure.assertHasCause("Resolving configuration 'internal' directly is not allowed")
-
-        where:
-        role                      | code
-        'consume or publish only' | 'queryOrResolveAllowed = false'
-        'bucket'                  | 'queryOrResolveAllowed = false; consumeOrPublishAllowed = false'
-
-    }
-
-    @Unroll("cannot resolve a configuration with role #role at configuration time")
-    def "cannot resolve a configuration which is for publishing only at configuration time"() {
-        given:
-        buildFile << """
-
-        configurations {
-            internal {
-                $code
-            }
-        }
-        dependencies {
-            internal files('foo.jar')
-        }
-
-        task checkState(dependsOn: configurations.internal.files) {
-        }
-
-        """
-
-        when:
-        fails 'checkState'
-
-        then:
-        failure.assertHasCause("Resolving configuration 'internal' directly is not allowed")
-
-        where:
-        role                      | code
-        'consume or publish only' | 'queryOrResolveAllowed = false'
-        'bucket'                  | 'queryOrResolveAllowed = false; consumeOrPublishAllowed = false'
-
-    }
-
-    @Unroll("cannot resolve a configuration with role #role using #method")
-    def "cannot resolve a configuration which is for publishing only"() {
-        given:
-        buildFile << """
-
-        configurations {
-            internal {
-                $role
-            }
-        }
-        dependencies {
-            internal files('foo.jar')
-        }
-
-        task checkState {
-            doLast {
-                configurations.internal.$method
-            }
-        }
-
-        """
-
-        when:
-        fails 'checkState'
-
-        then:
-        failure.assertHasCause("Resolving configuration 'internal' directly is not allowed")
-
-        where:
-        [method, role] << [
-            ['getResolvedConfiguration()', 'getBuildDependencies()', 'getIncoming().getFiles()', 'getIncoming().getResolutionResult()', 'getResolvedConfiguration()'],
-            ['queryOrResolveAllowed = false', 'queryOrResolveAllowed = false; consumeOrPublishAllowed = false']
-        ].combinations()
-    }
-
-    @Unroll("cannot add a dependency on a configuration role #role")
-    def "cannot add a dependency on a configuration not meant to be consumed or published"() {
-        given:
-        file('settings.gradle') << 'include "a", "b"'
-        buildFile << """
-        project(':a') {
-            configurations {
-                compile
-            }
-            dependencies {
-                compile project(path: ':b', configuration: 'internal')
-            }
-
-            task check {
-                doLast { configurations.compile.resolve() }
-            }
-        }
-        project(':b') {
-            configurations {
-                internal {
-                    $code
-                }
-            }
-        }
-
-        """
-
-        when:
-        fails 'a:check'
-
-        then:
-        failure.assertHasCause "Configuration 'internal' cannot be used in a project dependency"
-
-        where:
-        role                    | code
-        'query or resolve only' | 'consumeOrPublishAllowed = false'
-        'bucket'                | 'queryOrResolveAllowed = false; consumeOrPublishAllowed = false'
-    }
 }
